requirements.txt
openai>=1.83.0
python-dotenv>=1.0.0
pydantic>=2.0.0
